--- src/api/card.tsx
@@ -1,9 +1,0 @@
-import axios from "axios";
-
-export const getCategories = async() => {
-  const {data} = await axios({
-    method: "GET",
-    url: `http://192.168.1.33:8082/api/v1/cards/categories`
-  });
-  return data;
-}
