+++ src/app/register/page.tsx
@@ -3,11 +3,12 @@
 import { useSignupStore } from "@/store/signupStore";
 import { useRouter } from "next/navigation";
 import { useState } from "react";
-import SignupInput, { SignupInputProps } from "../login/components/SignupInput";
+import SignupInput, { SignupInputProps } from "./components/SignupInput";
 import { Register } from "@/api/auth";
+import { saveInterestCtg } from "@/api/category";
 
 const RegisterPage = () => {
-  const { newUser } = useSignupStore();
+  const { newUser, setNewUser } = useSignupStore();
   const [step, setStep] = useState(0);
   const router = useRouter();
 
@@ -24,7 +25,11 @@ const RegisterPage = () => {
       title: "phoneNum",
       descText: "휴대폰 번호를 입력해주세요.",
       inputText: "휴대폰 번호",
-      onNext: () => setStep(2),
+      // onNext: () => setStep(2),
+      onNext: () => {
+        console.log(newUser.rrn);
+        setStep(2);
+      },
       maxLen: 11,
       text: ""
     },
@@ -42,14 +47,13 @@ const RegisterPage = () => {
       inputText: "별명",
       onNext: async() => {
         try {
-          // const data = await Register(newUser.name, newUser.rrn, newUser.phoneNum, newUser.nickname, newUser.token);
-          // alert(data);
-          // router.push("/login");
+          const data = await Register(newUser.name, newUser.rrn, newUser.phoneNum, newUser.nickname, newUser.token);
+          setNewUser({...newUser, rrn: data});
           setStep(4);
         } catch (error) {
           console.error("Failed to validation code:", error);
-          // alert("회원가입에 실패했습니다.");
-          // window.location.reload();
+          alert("회원가입에 실패했습니다.");
+          window.location.reload();
         }
       },
       maxLen: 10,
@@ -59,8 +63,18 @@ const RegisterPage = () => {
       title: "cardBenefits",
       descText: "관심있는 카드 혜택을 5가지 선택해주세요.",
       inputText: "",
-      onNext: () => {
-        // 회원가입 api
+      onNext: async() => {
+        try {
+          const cardBenefitsAsNumbers = newUser.cardBenefits.map(Number);
+          console.log(newUser);
+          await saveInterestCtg(cardBenefitsAsNumbers, newUser.token);
+          alert("관심 카드혜택 등록 성공");
+          router.push("/login");
+        } catch (error) {
+          console.error("Failed to validation code:", error);
+          alert("관심 카드혜택 등록 실패");
+          // window.location.reload();
+        }
       },
       maxLen: 10,
       text: ""
